refactor: missed authorization owner type changes
